requirements.txt
flake8==3.8.2
h5py==2.10.0
ipython==6.5.0
joblib==0.13.2
jupyter==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
notebook==6.0.1
numpy==1.18.1
opencv-python==4.1.0.25
pandas==0.25.1
pickleshare==0.7.5
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Sphinx==3.0.4
tensorflow-gpu==1.14.0
tqdm==4.41.1
yellowbrick==1.1
